docs: Automatically generated library documentation (alumina-doc) (#15)
Another shameless ripoff of the Rust ecosystem, this time in the form of docs. `alumina-doc` is a tool that parses the stdlib and outputs HTML documentation.

It's very primitive at the moment, but it is at a stage where it's better than nothing. Also, it's baby steps towards a self-hosted compiler, since we are finally doing serious parsing in Alumina itself.