feat: Add multithreading support (#24)
Another in the series of ~Rust ripoffs~ Rust-inspired additions, this time threading.

This makes some opinionated decisions on the runtime environment when threads are present, so `threading` attribute is added to globally disable threading functionality in the standard library. Omitting this also makes `#[thread_local]` attribute a no-op, so it is not safe to link non-multithreaded Alumina code with multi-threaded code in general.