feat: Coerce the branches of an if expression to the type of the provided hint (#50)
The types of the two branches in an if expression currently have to be the same type with the exception of pointer mutability and divergence (if one branch diverges it's fine).

However, it can be useful to take higher-level coercions into account as well. This cannot be automatic, since there are an infinite number of types it could be coerced to, but if a type hint is present, it can be done.

This is mostly to support constructs where the two branches create a pointer to a specific type which is then coerced into a dyn pointer.